requirements.txt
Flask==2.0.2
python-dotenv==0.21.0
requests==2.28.1
twilio==8.5.0
gunicorn
ffmpeg-python
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.1
numpy==1.26.4
soundfile==0.12.1

# Google Gemini SDK
google-generativeai==0.8.3

# Cohere for sentiment analysis
cohere

# Text-to-Speech
gtts==2.4.0

# Additional audio processing
scipy==1.11.4

# Telegram Bot API
python-telegram-bot==20.7

# Additional Twilio features
twilio[voice]==8.5.0
